+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Changed
 
+- `PusPacket` trait: `user_data` now returns `&[u8]` instead of `Option<&[u8]>`. Empty user data
+  can simply be an empty slice.
 - Moved ECSS TC components from `tc` to `ecss.tc`.
 - Moved ECSS TM components from `tm` to `ecss.tm`.
 - Converted `PusTc` class to more specialized `PusTcCreator` and `PusTcReader`
@@ -23,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Added
 
+- `source_data` and `app_data` API provided for PUS TM and PUS TC reader classes. These simply
+  call `user_data` but are also in line with the PUS packet standard names for those fields.
 - Added new marker trait `IsPusTelemetry` implemented by `PusTmCreator` and `PusTmReader`.
 - Added new marker trait `IsPusTelecommand` implemented by `PusTcCreator` and `PusTcReader`.